requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
arrow==1.2.3
    # via -r requirements.in
certifi==2022.12.7
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via typer
colorama==0.4.6
    # via typer
commonmark==0.9.1
    # via rich
idna==3.4
    # via requests
loguru==0.6.0
    # via -r requirements.in
pydantic==1.10.2
    # via -r requirements.in
pygments==2.13.0
    # via rich
python-dateutil==2.8.2
    # via arrow
pyyaml==6.0
    # via -r requirements.in
requests==2.28.1
    # via -r requirements.in
rich==12.6.0
    # via typer
shellingham==1.5.0
    # via typer
six==1.16.0
    # via python-dateutil
typer[all]==0.7.0
    # via -r requirements.in
typing-extensions==4.4.0
    # via pydantic
urllib3==1.26.13
    # via requests
